restructured folder structure for the frontend
