--- src/frontend/Components/SideBar/CurrentState.tsx
@@ -1,9 +1,0 @@
-import Grid from '@material-ui/core/Grid';
-
-export const CurrentState = (): JSX.Element => {
-    return (
-        <Grid item>
-            Current State
-        </Grid>
-    );
-}
