--- src/frontend/Components/SideBar/HamburgerMenu.tsx
@@ -1,9 +1,0 @@
-import Grid from '@material-ui/core/Grid';
-
-export const HamburgerMenu = (): JSX.Element => {
-    return (
-        <Grid item>
-            Hamburger Menu
-        </Grid>
-    );
-}
