--- src/frontend/Components/SideBar/NextPrevButton.tsx
@@ -1,10 +1,0 @@
-import Grid from '@material-ui/core/Grid';
-import Button from '@material-ui/core/Button';
-
-export const NextPrevButton = (): JSX.Element => {
-    return (
-        <Grid item>
-            <Button variant='contained'>Test button</Button>
-        </Grid>
-    );
-}
