+++ src/frontend/Containers/MainContainer.tsx
@@ -1,14 +1,14 @@
 import Container from '@material-ui/core/Container';
 import Grid from '@material-ui/core/Grid';
-import { SideBarContainer } from './SideBarContainer';
-import { StateTreeContainer } from './StateTreeContainer';
+import { SnapshotContainer } from './SnapshotContainer';
+import { VisualContainer } from './VisualContainer';
 
 export const MainContainer = (): JSX.Element => {
     return (
         <Container>
             <Grid container>
-                <SideBarContainer />
-                <StateTreeContainer />
+                <SnapshotContainer />
+                <VisualContainer />
             </Grid>
         </Container>
     );
